Move listSecretVersionIds to the tester base's mock client
+ Add useful comments for future reference
