+++ jdisc-cloud-aws/src/main/java/ai/vespa/secret/aws/testutil/AsmSecretTesterBase.java
@@ -8,7 +8,17 @@ package ai.vespa.secret.aws.testutil;
 import ai.vespa.secret.model.Key;
 import ai.vespa.secret.model.SecretVersionState;
 import com.yahoo.vespa.athenz.api.AwsRole;
+import software.amazon.awssdk.awscore.exception.AwsServiceException;
+import software.amazon.awssdk.core.exception.SdkClientException;
 import software.amazon.awssdk.services.secretsmanager.SecretsManagerClient;
+import software.amazon.awssdk.services.secretsmanager.model.InternalServiceErrorException;
+import software.amazon.awssdk.services.secretsmanager.model.InvalidNextTokenException;
+import software.amazon.awssdk.services.secretsmanager.model.InvalidParameterException;
+import software.amazon.awssdk.services.secretsmanager.model.ListSecretVersionIdsRequest;
+import software.amazon.awssdk.services.secretsmanager.model.ListSecretVersionIdsResponse;
+import software.amazon.awssdk.services.secretsmanager.model.ResourceNotFoundException;
+import software.amazon.awssdk.services.secretsmanager.model.SecretVersionsListEntry;
+import software.amazon.awssdk.services.secretsmanager.model.SecretsManagerException;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -17,6 +27,11 @@ import java.util.Map;
 import java.util.function.Function;
 
 /**
+ * Base class for ASM reader and writer testers.
+ * Expected mapping from Key to AWS secret id must be provided by each test.
+ * This mapping is used when manually writing/reading secrets to/from the 'secrets' map,
+ * so that lookup will fail if the mapping does not match what the production code uses.
+ *
  * @author gjoranv
  */
 public class AsmSecretTesterBase {
@@ -49,6 +64,20 @@ public class AsmSecretTesterBase {
         protected MockSecretsManagerClient(AwsRole awsRole) {
             this.awsRole = awsRole;
             clients.add(this);
+        }
+
+        // Used by both reader and writer testers
+        @Override
+        public ListSecretVersionIdsResponse listSecretVersionIds(ListSecretVersionIdsRequest request) throws InvalidNextTokenException, ResourceNotFoundException, InternalServiceErrorException, InvalidParameterException, AwsServiceException, SdkClientException, SecretsManagerException {
+            return ListSecretVersionIdsResponse.builder()
+                    .name(request.secretId())
+                    .versions(secrets.getOrDefault(request.secretId(), List.of()).stream()
+                                      .map(version -> SecretVersionsListEntry.builder()
+                                              .versionId(version.version())
+                                              .versionStages(List.of(toAwsStage(version.state())))
+                                              .build())
+                                      .toList())
+                    .build();
         }
 
         @Override
